fix : 리뷰 기반 수정 (주문 취소 로직 리팩토링)
- 비관적 락 도입으로 동시 요청 시 데이터가 꼬이는 문제를 해결
- 모든 금액 변수를 long으로 맞추고, 패널티 계산 정책을 반올림으로 변경

+++ src/main/java/com/project/rare_x_back/config/TossPaymentClient.java
@@ -30,7 +30,7 @@ public class TossPaymentClient {
      */
     public void cancelPayment(
             String paymentKey,
-            int cancelAmount,
+            long cancelAmount,
             String reason
     ) {
